Refactor: Introduce an 'inventory' package to contain communications with the sim invenrtory and its associatred database
